extract GAL234 trait from facs data
Add two scripts to extract GAL2/3/4 promoter fused with YFP strain
trait from facs data. ‘GAL234prDoubleGradTrait’ treat the data as
bimodal, which is not reasonable as the histograms do not show two
modes. So we choose ‘GAL234prUni_DoubleGrad’ eventually, which treat
the data as unimodal, and only extract ONE mean from all the cells. (no
ON/OFF fraction)